+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.8.5] - 2021-04-30
+
+### Fixed
+- Adjusted to use path site instead advanced country
+- Adjusted the total value of the installments presented at checkout
 ## [3.8.4] - 2021-04-30
 
 ### Added